Fix: save hash correctly
