+++ backend/sql_app/Utils/crud.py
@@ -6,6 +6,15 @@ import json
 from email_validator import validate_email, EmailNotValidError
 
 from . import models, schemas
+from fastapi.security import OAuth2PasswordBearer
+from typing import Union, Annotated
+from datetime import timedelta
+from fastapi import Depends
+
+SECRET_KEY = "812d5458ae18565ac79bd58fe2903e1fc85171b1e36f78e74e24673a474dd9f5"
+ALGORITHM = "HS256"
+
+oauth2_scheme = OAuth2PasswordBearer(tokenUrl="login")
 
 def ValidateMail(email):
     try:
@@ -23,7 +32,7 @@ def create_user(db: Session, user: schemas.UserCreate):
         salt = bcrypt.gensalt(rounds=12)
         password_bytes = user.password.encode('utf-8') 
         hashed_password = bcrypt.hashpw(password_bytes,salt) # postpone due to some error
-        db_user = models.User(email=user.email, hashed_password=hashed_password, username=user.username)
+        db_user = models.User(email=user.email, hashed_password=hashed_password.decode('utf-8'), username=user.username)
         db.add(db_user)
         db.commit()
         db.refresh(db_user)
@@ -95,3 +104,37 @@ def create_wordpic(db:Session,wordpic:schemas.WordPictureBase):
     db.commit()
     db.refresh(new_pic)
     return new_pic
+
+
+def create_access_token(data: dict, expires_delta: Union[timedelta, None] = None):
+    to_encode = data.copy()
+    if expires_delta:
+        expire = datetime.now(timezone.utc) + expires_delta
+    else:
+        expire = datetime.now(timezone.utc) + timedelta(minutes=15)
+    to_encode.update({"exp": expire})
+    encoded_jwt = jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)
+    return encoded_jwt
+
+async def get_current_user(token: Annotated[str, Depends(oauth2_scheme)]):
+    credentials_exception = HTTPException(
+        status_code=status.HTTP_401_UNAUTHORIZED,
+        detail="Could not validate credentials",
+        headers={"WWW-Authenticate": "Bearer"},
+    )
+    try:
+        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
+        username: str = payload.get("sub")
+        if username is None:
+            raise credentials_exception
+        token_data = TokenData(username=username)
+    except JWTError:
+        raise credentials_exception
+    user = get_user(fake_users_db, username=token_data.username)
+    if user is None:
+        raise credentials_exception
+    return user
+
+def verify_password(plain_text:str, hashed:str):
+    name = bcrypt.checkpw(plain_text.encode('utf-8'),hashed)
+    return name
